Renaming modules and moving magetab2json functions
