+++ converter/dm2magetab.py
